more stuff

+++ server/api/models.py
@@ -3,13 +3,13 @@ from django.db import models
 import datetime
 from django.utils import timezone
 class User(AbstractUser):
-    balance = models.IntegerField(null=True, default=10000)
+    balance = models.FloatField(null=True, default=10000)
     courses = models.ManyToManyField('Course', related_name='users')
 
 class BalancePoint(models.Model):
     id = models.AutoField(primary_key=True)
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='balance_points')
-    balance = models.IntegerField()
+    balance = models.FloatField()
     timestamp = models.IntegerField()
 
 class Course(models.Model):
@@ -18,13 +18,13 @@ class Course(models.Model):
     id = models.AutoField(primary_key=True)
     course_code = models.CharField(max_length=10, unique=True)
     name = models.CharField(max_length=100, null=True) # accepts null but shouldn't be possible
-    price = models.IntegerField(null=True)
+    price = models.FloatField(null=True)
 
 
 class PricePoint(models.Model):
     id = models.AutoField(primary_key=True)
     course = models.ForeignKey(Course, on_delete=models.CASCADE, related_name='price_points')
-    price = models.IntegerField()
+    price = models.FloatField()
     # date = models.DateTimeField(auto_now_add=True) #
     date = models.DateTimeField(default=timezone.now)
     timestamp = models.IntegerField()
@@ -35,7 +35,7 @@ class Stock(models.Model):
     course = models.ForeignKey(Course, on_delete=models.CASCADE, related_name='stocks')
     amount = models.IntegerField()
 
-class Order(models.Model):
+class Order(models.Model): # To be deleted
     id = models.AutoField(primary_key=True)
     course = models.ForeignKey(Course, on_delete=models.CASCADE, related_name='orders', null=True)
     stock = models.ForeignKey(Stock, on_delete=models.CASCADE, related_name='orders', null=True)
